feat: remove u32 restriction

+++ common/rust/template/src/part_2.rs
@@ -1,5 +1,5 @@
 use aoc_common_rust::errors::AocError;
 
-pub fn solve(filename: &str) -> Result<u32, AocError> {
+pub fn solve(filename: &str) -> Result<u64, AocError> {
     Ok(42)
 }
